First commit, basic version of the game
Added the first version of the game.
Two commands, "check" and "flag" can be used in a Twitch chat to play
the game.
When a player checks a bomb, he dies and is unable to play for 60
seconds.
When the game is over, a new game is started after 30 seconds.